refactor: Put summary in own module
